Improve documentation (#65)
Improve documentation
Improve Instafeed demo
