+++ Sources/ASCollectionView/ASCollectionView.swift
@@ -542,7 +542,7 @@ public struct ASCollectionView<SectionID: Hashable>: UIViewControllerRepresentab
 	}
 }
 
-// MARK: OnReachedEnd support
+// MARK: OnReachedBoundary support
 
 extension ASCollectionView.Coordinator
 {
@@ -865,61 +865,5 @@ public extension ASCollectionView
 		this.layout = Layout(createCustomLayout: createCustomLayout, configureCustomLayout: configureCustomLayout)
 		return this
 	}
-
-	func shrinkToContentSize(isEnabled: Bool, _ contentSize: Binding<CGSize?>, dimensionToShrink: ShrinkDimension) -> some View
-	{
-		SelfSizingWrapper(self, isEnabled: isEnabled, contentSize: contentSize, shrinkDirection: dimensionToShrink)
-	}
 }
 
-protocol ContentSize
-{
-	var contentSize: Binding<CGSize?>? { get set }
-}
-
-// MARK: Helpers for self-sizing collection
-
-public enum ShrinkDimension
-{
-	case horizontal
-	case vertical
-
-	var shrinkVertical: Bool
-	{
-		self == .vertical
-	}
-
-	var shrinkHorizontal: Bool
-	{
-		self == .horizontal
-	}
-}
-
-struct SelfSizingWrapper<Content: View & ContentSize>: View
-{
-	var contentSize: Binding<CGSize?>
-	var content: Content
-	var shrinkDirection: ShrinkDimension
-	var isEnabled: Bool
-
-	init(_ content: Content, isEnabled: Bool, contentSize: Binding<CGSize?>, shrinkDirection: ShrinkDimension)
-	{
-		self.content = content
-		self.contentSize = contentSize
-		self.shrinkDirection = shrinkDirection
-		self.isEnabled = isEnabled
-
-		self.content.contentSize = contentSize
-	}
-
-	var body: some View
-	{
-		content
-			.frame(
-				idealWidth: isEnabled && shrinkDirection.shrinkHorizontal ? contentSize.wrappedValue?.width : nil,
-				maxWidth: isEnabled && shrinkDirection.shrinkHorizontal ? contentSize.wrappedValue?.width : nil,
-				idealHeight: isEnabled && shrinkDirection.shrinkVertical ? contentSize.wrappedValue?.height : nil,
-				maxHeight: isEnabled && shrinkDirection.shrinkVertical ? contentSize.wrappedValue?.height : nil,
-				alignment: .topLeading)
-	}
-}
