+++ Sources/ASCollectionView/EnvironmentKeys.swift
@@ -107,41 +107,51 @@ public extension EnvironmentValues
 
 public extension View
 {
+	/// Set whether to show scroll indicators for the ASCollectionView/ASTableView
 	func scrollIndicatorsEnabled(_ enabled: Bool) -> some View
 	{
 		environment(\.scrollIndicatorsEnabled, enabled)
 	}
 
+	/// Set insets for the ASCollectionView/ASTableView
 	func contentInsets(_ insets: UIEdgeInsets) -> some View
 	{
 		environment(\.contentInsets, insets)
 	}
 
+	/// Set whether the ASTableView should show separators
 	func tableViewSeparatorsEnabled(_ enabled: Bool) -> some View
 	{
 		environment(\.tableViewSeparatorsEnabled, enabled)
 	}
 
+	/// Set a closure that is called whenever the tableView is scrolled to the bottom.
+	/// This is useful to enable loading more data when scrolling to bottom
 	func onTableViewReachedBottom(_ onReachedBottom: @escaping (() -> Void)) -> some View
 	{
 		environment(\.tableViewOnReachedBottom, onReachedBottom)
 	}
-
+	
+	/// Set a closure that is called whenever the collectionView is scrolled to a boundary. eg. the bottom.
+	/// This is useful to enable loading more data when scrolling to bottom
 	func onCollectionViewReachedBoundary(_ onReachedBoundary: @escaping ((Boundary) -> Void)) -> some View
 	{
 		environment(\.collectionViewOnReachedBoundary, onReachedBoundary)
 	}
 
+	/// Set whether the ASCollectionView should always allow horizontal bounce
 	func alwaysBounceHorizontal(_ alwaysBounce: Bool = true) -> some View
 	{
 		environment(\.alwaysBounceHorizontal, alwaysBounce)
 	}
 
+	/// Set whether the ASCollectionView/ASTableView should always allow horizontal bounce
 	func alwaysBounceVertical(_ alwaysBounce: Bool = true) -> some View
 	{
 		environment(\.alwaysBounceVertical, alwaysBounce)
 	}
 
+	/// Set an initial scroll position for the ASCollectionView
 	func initialScrollPosition(_ scrollPosition: ASCollectionViewScrollPosition?) -> some View
 	{
 		environment(\.initialScrollPosition, scrollPosition)
